pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "lattice-enrichment"
version = "0.2.0"
authors = [
  { name="Lattice Team", email="team@lattice.dev" },
]
description = "A powerful tool for enriching CSV data using Large Language Models (LLM) on a row-by-row basis"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "pandas>=2.0.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langchain>=0.3.0",
    "tqdm>=4.65.0",
    "tenacity>=9.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "isort>=5.12",
    "mypy>=1.0",
    "pre-commit>=3.0",
]
vector = [
    "faiss-cpu>=1.7.0",
    "pymupdf>=1.23.0",
    "ipython>=8.0.0",
]
all = [
    "lattice-enrichment[dev,vector]"
]

[project.urls]
"Homepage" = "https://github.com/lattice-team/lattice-enrichment"
"Bug Reports" = "https://github.com/lattice-team/lattice-enrichment/issues"
"Source" = "https://github.com/lattice-team/lattice-enrichment"
"Documentation" = "https://github.com/lattice-team/lattice-enrichment#readme"

[project.scripts]
lattice-enrich = "lattice.cli:main"

[tool.hatch.version]
path = "lattice/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["lattice"]

[tool.hatch.build.targets.sdist]
include = [
    "lattice/",
    "tests/",
    "examples/",
    "docs/",
    "README.md",
    "pyproject.toml",
]

# Development tools configuration
[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pandas.*",
    "tqdm.*",
    "langchain.*",
    "langchain_openai.*",
    "langchain_core.*",
    "tenacity.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--disable-warnings",
    "-v"
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]